Update 1-enoch.md Chapters 1-13
Notes to consider:


I lowercased certain letters that came after a comma, since the paragraphs are formatted differently than the archive version.

Issues with numbers and letters mixed up still exists.

I'm leaving parentheses in, since they mark where a word was added (similar to how KJV italicizes words)

Chapter 10:9 says "Gabriel said to the Lord" in both the archive and this version, however in this context, the Lord is giving angels assignments." I need advice on whether I should leave it alone or not.

Chapter 13:6 had length length of days in this file, but "of days" was left out in the archive version. I think I took it out of this text. Should I put it back?